ai-multiagent-ticket-resolver-main-jasir115/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pytest
httpx
google-generativeai
streamlit
